Fix compatibility with new large VMTs
Some of the new VMT files for games like CSGO exceed the 1000 byte limit for data that is preloaded in the VPK tree. The rest of it is stored elsewhere in the archive. Before this, the application assumed the data for VMTs (usually a couple hundred bytes) was either all preloaded, or all located elsewhere in the file. Since the only time it's in both is if it's too large to be preloaded, I've only seen this show up with `hr_` textures in csgo.